+++ src/components/Review/index.tsx
@@ -1,10 +1,1 @@
-import { ReviewItem } from '~/components/Review/ReviewItem';
-import { ReviewWrapper } from '~/components/Review/ReviewWrapper';
-
-export const Review = Object.assign(
-  {},
-  {
-    Wrapper: ReviewWrapper,
-    Item: ReviewItem,
-  }
-);
+export { Review } from './Review';
